refactor CoursePartDialog to one component
